Handling missing teams in uploaded schedules
We had not provided any quality check on uploaded schedules. It is possible for schedules to have missing teams or teams that are not true FRC teams. We now detect missing teams in a driver's station and alert the user to adjust the spreadsheet. We also detect fake teams and assign them the barebones details, which the GA can adjust later.
Fixes #1