More work on the bracket
Including typekit fonts for bracket
Moving the async functions for fetching API data up and out of the useeffect so they can be used by other triggers, like when we navigate matches
Making it so that the bracket only shows when it's time
Adding more data to the bracket.
todo: adjust positions of the text blocks so they will be centered.