Adding filters and styling to the event list
Included two filters for the event list: one for time, and the other for type of event.

The time filter is a reductive filter, in that it reduces the possible events by that filter (past and future).

The type filter is additive, in that you can multi-select and it will add events to the dropdown.

I also included styling to the event list items: regular season events are black and white; offseason events connected to FMS are light blue and have •• wrapping the names, and offseason events not connected to FMS are light yellow and have •• wrapping the name.