Fix crasher in Playoffs and add support for Overtime 3
User identified a crasher on Playoff page, caused when reloading event when on the /allianceselection route. This is now fixed.
Closes #72
Users do not see which Alliance won the event. This is now indicated with a gold outline for the Finals matches.
Closes #73
Playoffs did not support 3 overtime matches, which are indicted in the rules. We have added another Finals match to support, and we are testing for the winner based on match 18 tiebreaker in the event of repeated ties.
Closes #71