Implementing process to keep event details up to date
Adding switch to enable behavior that keeps the app focused on the next match to be played. This is useful for the Emcee screen.
Adding switch
Adding timer
Adding function to fetch event details

Fixes a bug in the schedule loader that did not count the playoff matches as having been completed.
Updating cheat sheet metadata per feedback on Chief Delphi
Closes #205
Closes #206