Rich Text Editor
Adding RTE for the two team notes fields. Users can now add and edit rich text (with limitations) for these two fields.
The notes that appear on the Team Data screen allow many formatting options, including bullets, links, font size and type styling.
The notes that appear on teh Announce and Play-By-Play screens are limited to bold/italic/underline formatting.

This also changes how we handle the passage of time on the Team Data screen. Additional work is needed to reduce the number of redraws that are caused by these on-screen time-dependent displays, such as lastVisit and match timing (late, on-time etc).