Finer controls for loading teams
We had implemented a boolean control for ensuring that team loads don't conflict. We realized that if a user switched events during a load, the team list would never resolve.
Adds playoff support for OFFLINE matches (not playoff bracket yet). User can upload a Playoff match schedule after a Quals schedule, and then replace the Playoff schedule after each round completes.
Alliances are calculated based on Round One. Since substitutes are not allowed in Round 1, this should be good enough to know the Alliances, but not the Captains as the positions change.
We implemented backup team support, which may not persist through the tournament as you replace the teams. GAs can add backup teams as necessary.
Closes #311