OFFLINE support for Einstein
This adds support for running Einstein based on an excel export of Round 1. The user must enable new configuration options on the schedule page: Playoff-only OFFLINE Event and is World Champs BEFORE uploading the schedule. Once the schedule uploads, the user can then adjust the names of the Alliances and add the fourth member. Advancement through the playoff rounds is done with the Playoff bracket.
This update also includes a button to show the update history.
Closes #400
closes #399