Starting work on the FTC Local Server option
FTC says there will be a new local server in October 2025. This may introduce additional changes to the API for both the local and Cloud services. This sets the stage for those changes, and we can adjust as necessary in October.
THis introduces the FTC Server configuration options on the Setup Page. Users must enter the URL to the local server to access the data from it. Since there is no guarantee that the server will always have the same IP address, we need to offer the configuration.