Switching from statbotics/ftcScout direct calls to gatool API
This switches calls to statbotics and ftcScout to gatool API. This allows us to enable an unsecured endpoint for ftc local server and cheesy arena modes.

This also reduces the number of spurious 500 errors that will occur when loading an event that has not started. These were caused by calls to endpoints that returned 500 errors when there was no data. We can ensure that these endpoints query only when we know there should be data there.