Adding TBA support for loading events
This update brings support to load event details from TBA instead of FIRST APIs. gatool will automatically select this strategy when the user chooses an Offseason event that is not syncing with the FIRST APIs. This is especially useful for events that use Cheesy Arena, since most of those events do sync with TBA instead of FIRST.

This update also adds a new FTC Cheat Sheet to the cheatsheet page.