Aligning offseason demo teams with a specific event
FRC uses a range of offseason demo teams so that events can include new or extra teams in an offseason event. When we look for teamUpdates, we go by the number, but events often share the same teams.
This update ensures that offseason demo teams are keyed to specific events.
Closes #580