Schedule page updates
This refines how we process uploaded XLSX and CSV files for OFFLINE events. It excludes rows with no teams in the stations and removes an alert that was displaying incorrectly.

This also ensures that event details for events that we read from TBA are processed properly. The start and end date were not reading correctly, an some event details were getting lost.